did: remove external serializer

+++ semicongine/storage.nim
@@ -10,7 +10,6 @@ import std/times
 import ./core
 
 import ./thirdparty/db_connector/db_sqlite
-import ./thirdparty/vsbf/vsbf
 
 const STORAGE_NAME = Path("storage.db")
 const DEFAULT_KEY_VALUE_TABLE_NAME = "shelf"
@@ -114,35 +113,29 @@ proc ensureExists(worldName: string, table: string): DbConn =
     )
   )
 
-proc storeWorld*[T: object | tuple](
+proc storeWorld*[T](
     worldName: string, world: T, table = DEFAULT_WORLD_TABLE_NAME, deleteOld = false
 ) =
   let db = worldName.ensureExists(table)
   defer:
     db.close()
   let key = $(int(now().toTime().toUnixFloat() * 1000))
+  let stm = db.prepare(&"""INSERT INTO {table} VALUES(?, ?)""")
+  stm.bindParam(1, key)
+  stm.bindParam(2, $$world)
+  db.exec(stm)
+  stm.finalize()
+  if deleteOld:
+    db.exec(sql(&"""DELETE FROM {table} WHERE key <> ?"""), key)
 
-  var encoder = Encoder.init()
-  encoder.serializeRoot(world)
-
-  let s = db.prepare(&"""INSERT INTO {table} VALUES(?, ?)""")
-  s.bindParam(1, key)
-  s.bindParam(2, encoder.data)
-  db.exec(s)
-  s.finalize()
-  db.exec(sql(&"""DELETE FROM {table} WHERE key <> ?"""), key)
-
-proc loadWorld*[T: object | tuple](
-    worldName: string, table = DEFAULT_WORLD_TABLE_NAME
-): T =
+proc loadWorld*[T](worldName: string, table = DEFAULT_WORLD_TABLE_NAME): T =
   let db = worldName.ensureExists(table)
   defer:
     db.close()
   let dbResult =
     db.getValue(sql(&"""SELECT value FROM {table} ORDER BY key DESC LIMIT 1"""))
 
-  var decoder = Decoder.init(cast[seq[byte]](dbResult))
-  decoder.deserialize(T)
+  to[T](dbResult)
 
 proc listWorlds*(): seq[string] =
   let dir = Path(getDataDir()) / Path(AppName()) / Path(WORLD_DIR)
